backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.1
sqlalchemy==2.0.41
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
fastapi-users[sqlalchemy]==14.0.1
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0
crewai==0.148.0
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12
openai==1.73.0
faiss-cpu==1.11.0
numpy==1.26.4
google-auth==2.40.3
requests==2.32.3 